Fix production rubocop offenses

+++ lib/stradivari/table/builder/action_builder.rb
@@ -3,16 +3,22 @@ module Stradivari
     class Builder::ActionBuilder < Builder
       def self.render
         lambda do |object, _, opts|
-          actions = opts.fetch(:actions, [:edit, :delete])
+          actions = opts.fetch(:actions, %i[edit delete])
 
-          capture_haml do 
-            haml_concat link_to("<i class='fa fa-info'></i>".   html_safe, object,
-                     class: "btn btn-info btn-xs",    title: 'Show')   if actions.include?(:show)
-            haml_concat link_to("<i class='fa fa-edit'></i>".   html_safe, [:edit, object],
-                     class: "btn btn-primary btn-xs", title: 'Edit')   if actions.include?(:edit)
-            haml_concat link_to("<i class='fa fa-trash-o'></i>".html_safe, object,
-                     method: :delete, data: {confirm: "Do you want to remove this entity?"},
-                     class: "btn btn-danger btn-xs",  title: 'Delete') if actions.include?(:delete)
+          capture_haml do
+            if actions.include?(:show)
+              haml_concat link_to("<i class='fa fa-info'></i>".html_safe, object,
+                                  class: 'btn btn-info btn-xs',    title: 'Show')
+            end
+            if actions.include?(:edit)
+              haml_concat link_to("<i class='fa fa-edit'></i>".html_safe, [:edit, object],
+                                  class: 'btn btn-primary btn-xs', title: 'Edit')
+            end
+            if actions.include?(:delete)
+              haml_concat link_to("<i class='fa fa-trash-o'></i>".html_safe, object,
+                                  method: :delete, data: { confirm: 'Do you want to remove this entity?' },
+                                  class: 'btn btn-danger btn-xs', title: 'Delete')
+            end
           end
         end
       end
